fixed an issue causing a segfault: when creating a child environment, the type of the locals were being used instead of the type of the globals when loading

+++ compiler/codegen/scope.cpp
@@ -20,7 +20,7 @@ namespace codegen {
   }
 
   void GScope::finalize() {
-    for (int i = 0; i < functions.size(); i++) {
+    for (int i = 0; i < (int) functions.size(); i++) {
       auto function = functions[i];
       auto declaration = functionDeclarations[i];
       declaration->generateBody(function, this);
